+++ rush-plugins/rush-azure-storage-build-cache-plugin/src/AzureAuthenticationBase.ts
@@ -17,6 +17,7 @@ import { CredentialCache } from '@rushstack/rush-sdk';
 // See https://github.com/microsoft/rushstack/issues/3432
 import type { ICredentialCacheEntry } from '@rushstack/rush-sdk';
 import { PrintUtilities } from '@rushstack/terminal';
+import { AdoCodespacesAuthCredential } from './AdoCodespacesAuthCredential';
 
 /**
  * @public
@@ -79,7 +80,7 @@ export type AzureEnvironmentName = keyof typeof AzureAuthorityHosts;
 /**
  * @public
  */
-export type LoginFlowType = 'DeviceCode' | 'InteractiveBrowser';
+export type LoginFlowType = 'DeviceCode' | 'InteractiveBrowser' | 'AdoCodespacesAuth';
 
 /**
  * @public
@@ -88,6 +89,21 @@ export interface IAzureAuthenticationBaseOptions {
   azureEnvironment?: AzureEnvironmentName;
   credentialUpdateCommandForLogging?: string | undefined;
   loginFlow?: LoginFlowType;
+  /**
+   * A map to define the failover order for login flows. When a login flow fails to get a credential,
+   * the next login flow in the map will be attempted. If the login flow fails and there is no next
+   * login flow, the error will be thrown.
+   *
+   * @defaultValue
+   * ```json
+   * {
+   *   "AdoCodespacesAuth": "InteractiveBrowser",
+   *   "InteractiveBrowser": "DeviceCode",
+   *   "DeviceCode": null
+   * }
+   * ```
+   */
+  loginFlowFailover?: Record<LoginFlowType, LoginFlowType | undefined>;
 }
 
 /**
@@ -112,6 +128,7 @@ export abstract class AzureAuthenticationBase {
 
   protected readonly _azureEnvironment: AzureEnvironmentName;
   protected readonly _loginFlow: LoginFlowType;
+  protected readonly _failoverOrder: Record<LoginFlowType, LoginFlowType | undefined>;
 
   private __credentialCacheId: string | undefined;
   private get _credentialCacheId(): string {
@@ -132,6 +149,11 @@ export abstract class AzureAuthenticationBase {
     this._azureEnvironment = options.azureEnvironment || 'AzurePublicCloud';
     this._credentialUpdateCommandForLogging = options.credentialUpdateCommandForLogging;
     this._loginFlow = options.loginFlow || 'DeviceCode';
+    this._failoverOrder = options.loginFlowFailover || {
+      AdoCodespacesAuth: 'InteractiveBrowser',
+      InteractiveBrowser: 'DeviceCode',
+      DeviceCode: undefined
+    };
   }
 
   public async updateCachedCredentialAsync(terminal: ITerminal, credential: string): Promise<void> {
@@ -274,6 +296,10 @@ export abstract class AzureAuthenticationBase {
     };
 
     switch (loginFlow) {
+      case 'AdoCodespacesAuth': {
+        tokenCredential = new AdoCodespacesAuthCredential();
+        break;
+      }
       case 'InteractiveBrowser': {
         tokenCredential = new InteractiveBrowserCredential(interactiveCredentialOptions);
         break;
@@ -292,6 +318,17 @@ export abstract class AzureAuthenticationBase {
       }
     }
 
-    return await this._getCredentialFromTokenAsync(terminal, tokenCredential);
+    try {
+      return await this._getCredentialFromTokenAsync(terminal, tokenCredential);
+    } catch (error) {
+      terminal.writeVerbose(`Failed to get credentials with ${loginFlow}: ${error}`);
+      const fallbackFlow: LoginFlowType | undefined = this._failoverOrder[loginFlow];
+      if (fallbackFlow) {
+        terminal.writeVerbose(`Falling back to ${fallbackFlow} login flow`);
+        return this._getCredentialAsync(terminal, fallbackFlow);
+      } else {
+        throw error;
+      }
+    }
   }
 }